projects/vibecodingplatform-mvp/backend/requirements.txt
# Backend Dependencies for Vibecoding Platform MVP

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Environment Variables
python-dotenv==1.0.0

# Core gpt-engineer dependencies (should already be installed in main project)
# If running standalone, uncomment these:
# openai>=1.0.0
# langchain>=0.1.0
# langchain-openai>=0.0.5
# langchain-anthropic>=0.1.0
# tiktoken>=0.5.0

